+++ nano/node/nodeconfig.hpp
@@ -107,9 +107,9 @@ public:
 	nano::lmdb_config lmdb_config;
 	nano::frontiers_confirmation_mode frontiers_confirmation{ nano::frontiers_confirmation_mode::automatic };
 	/** Number of accounts per second to process when doing backlog population scan */
-	uint64_t backlog_scan_rate{ 10 * 1000 };
+	unsigned backlog_scan_rate{ 10 * 1000 };
 	/** Number of times per second to run backlog population batches. Number of accounts per single batch is `backlog_scan_rate / backlog_scan_frequency` */
-	uint backlog_scan_frequency{ 10 };
+	unsigned backlog_scan_frequency{ 10 };
 
 public:
 	std::string serialize_frontiers_confirmation (nano::frontiers_confirmation_mode) const;
